+++ sc-kpm/sc-ui/uiTypes.h
@@ -17,11 +17,22 @@ extern "C"
 #include <map>
 #include <list>
 #include <vector>
+#include <unordered_set>
 #include <assert.h>
 #include <stdint.h>
 
 typedef std::string String;
 typedef std::stringstream StringStream;
+
+struct _sc_addr_hash
+{
+  auto operator()(const sc_addr & addr) const -> size_t
+  {
+    return SC_ADDR_LOCAL_TO_INT(addr);
+  }
+};
+
+#include <nlohmann/json.hpp>
 
 typedef std::list<sc_addr> tScAddrList;
 typedef std::vector<sc_addr> tScAddrVector;
@@ -30,6 +41,9 @@ typedef std::map<sc_addr, sc_addr> tScAddrToScAddrMap;
 typedef std::pair<sc_addr, sc_addr> tScAddrPair;
 typedef std::list<tScAddrPair> tScAddrPairList;
 typedef std::map<String, String> tStringStringMap;
+typedef std::unordered_set<sc_addr, _sc_addr_hash> tScAddrSet;
+
+typedef nlohmann::json sc_json;
 
 // --- operators ---
 bool operator<(const sc_addr & addr1, const sc_addr & addr2);