+++ docs/changelog.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Append generated elements by `SCs-helper` into requested output structure
 - Oriented sets representation (<a, ..., b>) in SCs
 
+### Changed
+
+- Add language parameter for sc2scs-json-translator to get elements identifiers
+- Append to sc2scs-json elements main/system identifiers
+- Replace row strings by nlohmann-json in sc2scs-json-translator
+
 ### Fixed
 
 - Add curl to installation script